Added fzf.vim submodule
This commit adds the 'fzf.vim' submodule to the repository. This submodule is used to enhance the functionality of the project.

Submodules are a great way to include external repositories as dependencies, ensuring that we can easily update them as needed.

The submodule was added using the following command: git submodule add git@github.com:bwambale03/fzf.vim.git fzf.vim